Split into IS and SA for BAT and WAT
Added further divisions of the masking of MRIs, for BAT and WAT there are now two subdivisions, IS and SA. This change was made following the request by Mya via email: Dear Daniel and Ariane,

Sorry for the delay in replying your email.
Ivy has checked the updated program. The naming of supra-clavicular and axillary fat depot (TSA) are reflected correctly in the program interface as well as in the excel file.
Just for Inter-scapular, we can only select min and max FF for total IS depot. As with the TSA, we will also need FF of BAT and WAT within this IS depot. Therefore, we would need 2 more bars to select FFs for BAT and WAT for IS. You may then name the bars like;
Bar 1: SA BAT min and SA BAT max
Bar 2: SA WAT min and SA WAT max
Bar 3: TSA min and TSA max
Bar 4: IS BAT min and IS BAT max
Bar 5: IS WAT min and IS WAT max
Bar 6: TIS min and TIS max
And all these are to reflect in the excel file as well.
I am so very sorry to ask more of your help. Please take your time.